+++ typescript/infra/config/environments/mainnet3/owners.ts
@@ -128,16 +128,20 @@ export const icas: Partial<
 
   // Oct 16, 2024 batch
   // ----------------------------------------------------------
-  immutablezkevm: '0x8483e1480B62cB9f0aCecEbF42469b9f4013577a',
-  rari: '0x1124D54E989570A798769E534eAFbE1444f40AF6',
-  rootstock: '0x69350aeA98c5195F2c3cC6E6A065d0d8B12F659A',
-  alephzeroevm: '0x004a4C2e4Cd4F5Bd564fe0A6Ab2Da56258aE576f',
-  chiliz: '0xb52D281aD2BA9761c16f400d755837493e2baDB7',
-  lumia: '0x418E10Ac9e0b84022d0636228d05bc74172e0e41',
-  superposition: '0x34b57ff8fBA8da0cFdA795CC0F874FfaB14B1DE9',
-  flow: '0xf48377f8A3ddA7AAD7C2460C81d939434c829b45',
-  metall2: '0x2f1b1B0Fb7652E621316460f6c3b019F61d8dC9a',
-  polynomial: '0xC20eFa1e5A378af9233e9b24515eb3408d43f900',
+  // lumia: '0x418E10Ac9e0b84022d0636228d05bc74172e0e41',
+
+  // // Oct 30, 2024 batch
+  // // ----------------------------------------------------------
+  // apechain: '0xe68b0aB6BB8c11D855556A5d3539524f6DB3bdc6',
+  // arbitrumnova: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // b3: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // fantom: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // gravity: '0x3104ADE26e21AEbdB325321433541DfE8B5dCF23',
+  // harmony: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // kaia: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // morph: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // orderly: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
+  // snaxchain: '0x8965d9f19336EB4e910d5f1B9070205FdBee6837',
 } as const;
 
 export const DEPLOYER = '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba';
@@ -166,3 +170,30 @@ export const ethereumChainOwners: ChainMap<OwnableConfig> = Object.fromEntries(
     ];
   }),
 );
+
+export const chainOwners: ChainMap<OwnableConfig> = {
+  ...ethereumChainOwners,
+  solanamainnet: {
+    // Squads multisig
+    owner: 'BNGDJ1h9brgt6FFVd8No1TVAH48Fp44d7jkuydr1URwJ',
+  },
+  eclipsemainnet: {
+    owner: 'B32qTbw8iDcVdagMAMBfDFCef88N1KtfRG5QoPVzbj5K',
+  },
+  injective: {
+    // Native multisig
+    owner: 'inj1ac6qpt57vhtfzdecd2an052elwgenwtxcn9chl',
+  },
+  neutron: {
+    // Da0Da0 multisig
+    owner: 'neutron1fqf5mprg3f5hytvzp3t7spmsum6rjrw80mq8zgkc0h6rxga0dtzqws3uu7',
+  },
+  // We intentionally cause issues if these were to be used, but satisfy the types
+  // and ensure there's an entry for each supported chain.
+  stride: {
+    owner: 'n/a - nothing owned here',
+  },
+  osmosis: {
+    owner: 'n/a - nothing owned here',
+  },
+};