+++ typescript/infra/scripts/check/check-utils.ts
@@ -24,6 +24,7 @@ import { Contexts } from '../../config/contexts.js';
 import { DEPLOYER } from '../../config/environments/mainnet3/owners.js';
 import { getWarpAddresses } from '../../config/registry.js';
 import { getWarpConfig } from '../../config/warp.js';
+import { chainsToSkip } from '../../src/config/chain.js';
 import { DeployEnvironment } from '../../src/config/environment.js';
 import { HyperlaneAppGovernor } from '../../src/govern/HyperlaneAppGovernor.js';
 import { HyperlaneCoreGovernor } from '../../src/govern/HyperlaneCoreGovernor.js';
@@ -116,6 +117,7 @@ export async function getGovernor(
   );
 
   if (module === Modules.CORE) {
+    chainsToSkip.forEach((chain) => delete envConfig.core[chain]);
     const checker = new HyperlaneCoreChecker(
       multiProvider,
       core,
@@ -139,6 +141,7 @@ export async function getGovernor(
     );
     governor = new ProxiedRouterGovernor(checker);
   } else if (module === Modules.HAAS) {
+    chainsToSkip.forEach((chain) => delete routerConfig[chain]);
     const icaChecker = new InterchainAccountChecker(
       multiProvider,
       ica,
@@ -147,6 +150,7 @@ export async function getGovernor(
         (chain, _): _ is InterchainAccountConfig => !!icaChainAddresses[chain],
       ),
     );
+    chainsToSkip.forEach((chain) => delete envConfig.core[chain]);
     const coreChecker = new HyperlaneCoreChecker(
       multiProvider,
       core,