+++ typescript/infra/src/config/chain.ts
@@ -16,6 +16,20 @@ import { inCIMode } from '../utils/utils.js';
 
 import { DeployEnvironment } from './environment.js';
 
+// Temporarily skip some chains
+export const chainsToSkip: ChainName[] = [
+  // TODO: remove once zksync PR is merged into main
+  // mainnets
+  'zksync',
+  'zeronetwork',
+  // testnets
+  'abstracttestnet',
+  'treasuretopaz',
+
+  // Oct 16 batch
+  'lumia',
+];
+
 export const defaultRetry: ProviderRetryOptions = {
   maxRetries: 6,
   baseRetryDelayMs: 50,