+++ typescript/sdk/src/consts/multisigIsm.ts
@@ -3,13 +3,19 @@ import { ChainMap } from '../types.js';
 
 // TODO: consider migrating these to the registry too
 export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
-  alephzeroevm: {
-    threshold: 2,
-    validators: [
-      '0xcae8fab142adc4e434bb7409e40dd932cc3851aa',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  abstracttestnet: {
+    threshold: 1,
+    validators: ['0x7655bc4c9802bfcb3132b8822155b60a4fbbce3e'],
+  },
+
+  alephzeroevmmainnet: {
+    threshold: 1,
+    validators: ['0x33f20e6e775747d60301c6ea1c50e51f0389740c'],
+  },
+
+  alephzeroevmtestnet: {
+    threshold: 1,
+    validators: ['0x556cd94bcb6e5773e8df75e7eb3f91909d266a26'],
   },
 
   alfajores: {
@@ -64,9 +70,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0x09fabfbca0b8bf042e2a1161ee5010d147b0f603'],
   },
 
-  arcadiatestnet: {
+  arcadiatestnet2: {
     threshold: 1,
-    validators: ['0x7ce5973d3f22971546efb86f5a0417c1248e92f5'],
+    validators: ['0xd39cd388ce3f616bc81be6dd3ec9348d7cdf4dff'],
   },
 
   astar: {
@@ -206,13 +212,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  chiliz: {
-    threshold: 2,
-    validators: [
-      '0x82d024f453b1a3f3f6606226f06b038da27596f3',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  chilizmainnet: {
+    threshold: 1,
+    validators: ['0x7403e5d58b48b0f5f715d9c78fbc581f01a625cb'],
   },
 
   citreatestnet: {
@@ -321,6 +323,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
+  flame: {
+    threshold: 1,
+    validators: ['0x1fa928ce884fa16357d4b8866e096392d4d81f43'],
+  },
+
   flare: {
     threshold: 2,
     validators: [
@@ -330,13 +337,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  flow: {
-    threshold: 2,
-    validators: [
-      '0x3aee1090318e9c54d1d23194dcd0f2bee00ddc97',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  flowmainnet: {
+    threshold: 1,
+    validators: ['0xe132235c958ca1f3f24d772e5970dd58da4c0f6e'],
   },
 
   formtestnet: {
@@ -409,13 +412,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0xea673a92a23ca319b9d85cc16b248645cd5158da'],
   },
 
-  immutablezkevm: {
-    threshold: 2,
-    validators: [
-      '0xa787c2952a4d22f776ee6e87e828e6f75de24330',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  immutablezkevmmainnet: {
+    threshold: 1,
+    validators: ['0xbdda85b19a5efbe09e52a32db1a072f043dd66da'],
   },
 
   inevm: {
@@ -434,6 +433,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x6B1d09A97b813D53e9D4b7523DA36604C0B52242', // caldera
       '0x9e551b6694bbd295d7d6e6a2540c7d41ce70a3b9', // imperator
     ],
+  },
+
+  inksepolia: {
+    threshold: 1,
+    validators: ['0xe61c846aee275070207fcbf43674eb254f06097a'],
   },
 
   kaia: {
@@ -524,13 +528,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  metall2: {
-    threshold: 2,
-    validators: [
-      '0x1b000e1e1f0a032ed382c6d69a2d58f6fe773c09',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  metal: {
+    threshold: 1,
+    validators: ['0xd9f7f1a05826197a93df51e86cefb41dfbfb896a'],
   },
 
   metis: {
@@ -678,13 +678,14 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  polynomial: {
-    threshold: 2,
-    validators: [
-      '0xa63ad0891e921ad5947d57e05831fabb9816eca7',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  polynomialfi: {
+    threshold: 1,
+    validators: ['0x23d348c2d365040e56f3fee07e6897122915f513'],
+  },
+
+  prom: {
+    threshold: 1,
+    validators: ['0xb0c4042b7c9a95345be8913f4cdbf4043b923d98'],
   },
 
   proofofplay: {
@@ -696,13 +697,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  rari: {
-    threshold: 2,
-    validators: [
-      '0x989d6862e09de21337078efbd86843a3eb1133e3',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  rarichain: {
+    threshold: 1,
+    validators: ['0xeac012df7530720dd7d6f9b727e4fe39807d1516'],
   },
 
   real: {
@@ -724,13 +721,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     ],
   },
 
-  rootstock: {
-    threshold: 2,
-    validators: [
-      '0xcb8e3a72cf427feff27416d0e2ec375a052eaaee',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  rootstockmainnet: {
+    threshold: 1,
+    validators: ['0x8675eb603d62ab64e3efe90df914e555966e04ac'],
   },
 
   sanko: {
@@ -854,13 +847,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     validators: ['0xf58f6e30aabba34e8dd7f79b3168507192e2cc9b'],
   },
 
-  superposition: {
-    threshold: 2,
-    validators: [
-      '0x5978d0e6afa9270ddb87cff43a8fa7a763a5dfc4',
-      '0xCF0211faFBb91FD9D06D7E306B30032DC3A1934f', // merkly
-      '0x4f977a59fdc2d9e39f6d780a84d5b4add1495a36', // mitosis
-    ],
+  superpositionmainnet: {
+    threshold: 1,
+    validators: ['0x3f489acdd341c6b4dd86293fa2cc5ecc8ccf4f84'],
   },
 
   superpositiontestnet: {
@@ -885,6 +874,11 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0xcf0211fafbb91fd9d06d7e306b30032dc3a1934f', // merkly
       '0xe271ef9a6e312540f099a378865432fa73f26689', // tangle
     ],
+  },
+
+  treasuretopaz: {
+    threshold: 1,
+    validators: ['0x9750849beda0a7870462d4685f953fe39033a5ae'],
   },
 
   unichaintestnet: {