+++ typescript/sdk/src/gas/token-prices.ts
@@ -1,21 +1,14 @@
-import CoinGecko from 'coingecko-api';
-
-import { rootLogger, sleep } from '@hyperlane-xyz/utils';
+import { objKeys, rootLogger, sleep } from '@hyperlane-xyz/utils';
 
 import { ChainMetadata } from '../metadata/chainMetadataTypes.js';
 import { ChainMap, ChainName } from '../types.js';
+
+const COINGECKO_PRICE_API = 'https://api.coingecko.com/api/v3/simple/price';
 
 export interface TokenPriceGetter {
   getTokenPrice(chain: ChainName): Promise<number>;
   getTokenExchangeRate(base: ChainName, quote: ChainName): Promise<number>;
 }
-
-export type CoinGeckoInterface = Pick<CoinGecko, 'simple'>;
-export type CoinGeckoSimpleInterface = CoinGecko['simple'];
-export type CoinGeckoSimplePriceParams = Parameters<
-  CoinGeckoSimpleInterface['price']
->[0];
-export type CoinGeckoResponse = ReturnType<CoinGeckoSimpleInterface['price']>;
 
 type TokenPriceCacheEntry = {
   price: number;
@@ -66,35 +59,26 @@ class TokenPriceCache {
 }
 
 export class CoinGeckoTokenPriceGetter implements TokenPriceGetter {
-  protected coinGecko: CoinGeckoInterface;
   protected cache: TokenPriceCache;
+  protected apiKey?: string;
   protected sleepMsBetweenRequests: number;
   protected metadata: ChainMap<ChainMetadata>;
 
-  constructor(
-    coinGecko: CoinGeckoInterface,
-    chainMetadata: ChainMap<ChainMetadata>,
-    expirySeconds?: number,
+  constructor({
+    chainMetadata,
+    apiKey,
+    expirySeconds,
     sleepMsBetweenRequests = 5000,
-  ) {
-    this.coinGecko = coinGecko;
+  }: {
+    chainMetadata: ChainMap<ChainMetadata>;
+    apiKey?: string;
+    expirySeconds?: number;
+    sleepMsBetweenRequests?: number;
+  }) {
+    this.apiKey = apiKey;
     this.cache = new TokenPriceCache(expirySeconds);
     this.metadata = chainMetadata;
     this.sleepMsBetweenRequests = sleepMsBetweenRequests;
-  }
-
-  static withDefaultCoinGecko(
-    chainMetadata: ChainMap<ChainMetadata>,
-    expirySeconds?: number,
-    sleepMsBetweenRequests = 5000,
-  ): CoinGeckoTokenPriceGetter {
-    const coinGecko = new CoinGecko();
-    return new CoinGeckoTokenPriceGetter(
-      coinGecko,
-      chainMetadata,
-      expirySeconds,
-      sleepMsBetweenRequests,
-    );
   }
 
   async getTokenPrice(
@@ -103,6 +87,15 @@ export class CoinGeckoTokenPriceGetter implements TokenPriceGetter {
   ): Promise<number> {
     const [price] = await this.getTokenPrices([chain], currency);
     return price;
+  }
+
+  async getAllTokenPrices(currency: string = 'usd'): Promise<ChainMap<number>> {
+    const chains = objKeys(this.metadata);
+    const prices = await this.getTokenPrices(chains, currency);
+    return chains.reduce(
+      (agg, chain, i) => ({ ...agg, [chain]: prices[i] }),
+      {},
+    );
   }
 
   async getTokenExchangeRate(
@@ -153,25 +146,35 @@ export class CoinGeckoTokenPriceGetter implements TokenPriceGetter {
     await sleep(this.sleepMsBetweenRequests);
 
     if (toQuery.length > 0) {
-      let response: any;
       try {
-        response = await this.coinGecko.simple.price({
-          ids: toQuery,
-          vs_currencies: [currency],
-        });
-
-        if (response.success === true) {
-          const prices = toQuery.map((id) => response.data[id][currency]);
-          toQuery.map((id, i) => this.cache.put(id, prices[i]));
-        } else {
-          rootLogger.warn('Failed to query token prices', response.message);
-          return undefined;
-        }
+        const prices = await this.fetchPriceData(toQuery, currency);
+        prices.forEach((price, i) => this.cache.put(toQuery[i], price));
       } catch (e) {
         rootLogger.warn('Error when querying token prices', e);
         return undefined;
       }
     }
     return ids.map((id) => this.cache.fetch(id));
+  }
+
+  public async fetchPriceData(
+    ids: string[],
+    currency: string,
+  ): Promise<number[]> {
+    let url = `${COINGECKO_PRICE_API}?ids=${Object.entries(ids).join(
+      ',',
+    )}&vs_currencies=${currency}`;
+    if (this.apiKey) {
+      url += `&x-cg-pro-api-key=${this.apiKey}`;
+    }
+
+    const resp = await fetch(url);
+    const idPrices = await resp.json();
+
+    return ids.map((id) => {
+      const price = idPrices[id]?.[currency];
+      if (!price) throw new Error(`No price found for ${id}`);
+      return Number(price);
+    });
   }
 }